requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
filelock==3.18.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
narwhals==1.38.2
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
packaging==24.2
pandas==2.2.3
pillow==11.2.1
protobuf==6.30.2
pyarrow==20.0.0
pycparser==2.22
pydeck==0.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.24.0
setuptools==80.4.0
six==1.17.0
smmap==5.0.2
sounddevice==0.5.1
streamlit==1.45.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
torch==2.7.0
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
